refactor: 쿠키 관련 기능을 CookieService 클래스로 리팩토링
